+++ resources/templates/js/webview.sidebar.js
@@ -87,12 +87,7 @@ new Promise((resolve) => {
     vscode.window.registerWebviewViewProvider(
       viewId,
       createWebviewViewProvider(viewId, resolve)),
-    {
-      dispose: () => {
-        macros.window.releaseWebviewId(viewId);
-        vscode.commands.executeCommand('setContext', `${viewId}.show`, false);
-      }
-    },
+    { dispose: () => vscode.commands.executeCommand('setContext', `${viewId}.show`, false) },
   );
 
   vscode.commands.executeCommand('setContext', `${viewId}.show`, true);